ai-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pillow==11.3.0
numpy==2.3.2
openai>=1.40.0,<2.0.0
